Final refactor
