+++ concrete/src/Attribute/Context/Context.php
@@ -4,7 +4,6 @@ namespace Concrete\Core\Attribute\Context;
 use Concrete\Core\Entity\Attribute\Key\Key;
 use Concrete\Core\Entity\Attribute\Value\AbstractValue;
 use Concrete\Core\Filesystem\TemplateLocator;
-use Concrete\Core\Form\Group\ViewInterface;
 
 abstract class Context implements ContextInterface
 {
