+++ concrete/src/Attribute/Form/RendererBuilder.php
@@ -1,7 +1,7 @@
 <?php
 namespace Concrete\Core\Attribute\Form;
 
-use Concrete\Core\Attribute\Form\Control\View;
+use Concrete\Core\Attribute\Form\Control\View\View;
 use Concrete\Core\Entity\Attribute\Key\Key;
 
 class RendererBuilder
