+++ concrete/src/Express/Form/Control/View/View.php
@@ -2,12 +2,10 @@
 namespace Concrete\Core\Express\Form\Control\View;
 
 use Concrete\Core\Entity\Express\Control\Control;
-use Concrete\Core\Entity\Express\Form;
-use Concrete\Core\Filesystem\TemplateLocator;
 use Concrete\Core\Express\Form\Context\ContextInterface;
-use Concrete\Core\Form\Group\ControlView;
+use Concrete\Core\Form\Control\FormView as BaseFormView;
 
-abstract class View extends ControlView
+abstract class View extends BaseFormView
 {
 
     public function __construct(ContextInterface $context, Control $control)
