+++ concrete/src/Form/Context/ContextInterface.php
@@ -1,9 +1,7 @@
 <?php
 namespace Concrete\Core\Form\Context;
 
-use Concrete\Core\Attribute\Form\Control\View;
 use Concrete\Core\Filesystem\TemplateLocator;
-use Concrete\Core\Form\Group\ViewInterface;
 
 interface ContextInterface
 {
