+++ concrete/src/Form/Control/ControlInterface.php
@@ -2,7 +2,6 @@
 namespace Concrete\Core\Form\Control;
 
 use Concrete\Core\Form\Context\ContextInterface;
-use Concrete\Core\Form\Group\ViewInterface as GroupViewInterface;
 
 interface ControlInterface
 {
