+++ concrete/src/Form/Control/FormViewInterface.php
@@ -1,0 +1,10 @@
+<?php
+namespace Concrete\Core\Form\Control;
+
+interface FormViewInterface extends ViewInterface
+{
+    function isRequired();
+    function getLabel();
+    function setLabel($label);
+    function supportsLabel();
+}
