--- concrete/src/Form/Group/ControlViewInterface.php
@@ -1,12 +1,0 @@
-<?php
-namespace Concrete\Core\Form\Group;
-
-use Concrete\Core\Form\Control\ControlInterface;
-
-interface ControlViewInterface extends ViewInterface
-{
-    function isRequired();
-    function getLabel();
-    function setLabel($label);
-    function supportsLabel();
-}
